Update github-etiquette.md
added a number of details and wording changes.

note that in almost all instances where epic or release was used on its own, I changed it to "epic/release" to avoid confusion. 

note: now state that a project not only has its own github repo and zenhub ws, but also, its own folder on the hpc, and at least one assigned project lead

I sort of hope we can ultimately move away from the epic/release combo term, but for now, given that is how zenhub uses the term epic for what i think is a release, and because we cant just abandon the term release (it is to common of a term in software dev and agile) we are stuck. 

modified to say that at the start of the project all epics/releases need to be stated, but, only the majority of issues (not all) and only for the first epic/release need to be added to the repo. Users are welcome to add additional issues for future epics/releases as they see fit, but the priority is on the first epic/release, and the expectation is NOT to write down all issues at the start, just the majority. 

changed the steps for publishing a release to not make modfications to dev readme first. instead, the feature of interest branch should also contain an edited readme, that way, when that feature is merged to dev, the readme is updated as well. and, if rejected, we are not left with readme changes without a corresponding change in actual code, data, etc.

additionally, prs needed for merging to dev as well, to reach consensus. 

Added a number of other little things, but the above were notable.

keeping the sprint stuff for now, but, we may consider dropping and then only bringing back in the future if really want to commit to it. I.e. we should at least be able to regularly stick to such a protocol for current projects if we want to keep this as the "standard". Otherwise, if I (and others) continue to ignore the sprint workflow standards, then members may question the legitamacy of other things on the wiki. 

ditto for specific details of the sprint. E.g., time estimates. We may ultimately decide to keep a stripped-down and simple version of the sprint workflow (or get rid of it altogether). For now, all details are retained, but we should think critically and decide on a plan.

if keeping the sprint, we should bring back the "done" pipeline in the zenhub workspaces to facilitate review.

i