Update naming-conventions.md
updated rule for original project naming to require all initial projects to be appended with -alpha, given that it is often not clear if a spin-off will occur or not, and one cannot go back later to rename an original project. 

Given that future spin-offs of a project are common, yet sometimes hard to predict, the original project should always be named the "-alpha" project (even if no spin-off is initially planned/envisioned) and spin-offs should use an informative modifier.