Update zotero.md
@jessb0t  I really liked all of your edits, they made what I was trying to say much clearer. Thank you!! 

The only change I'm rejecting is the first one, where an extra comma was added: "...easy for users to add properly formatted, in-text citations".  My view is that "in-text" is acting as part of the "citation" noun ("in-text citation"), and not as a second adjective (in which case it would be "properly formatted, in-text, citations"). Since there's only one adjective, no comma needed.