Update social-context-alpha.md
@emach039 @jessb0t Great work here; thank you! The text is very nicely written and spot-on!!

Note: I cut the links provided, just because I felt like they might not be the best in this case. Perhaps we can discuss at a future lab meeting the kind of links to use? That said, the one for social anxiety was great, but I didn't want to leave just one link, so removed all for now.