testing whether changing branch fixes issue
Right now, as the whos-who.md is written in whos-who-brandon, the picture for brandon does not show up. I think it's due to the branch specified in the url to the picture. If changing the branch from main to whos-who-brandon fixes the issue, then once this branch is merged the picture should show up with main specified.